Move swap_struct_t to its own file.
This way bswap32() and friends are already defined.  Otherwise there are
errors.
